Segmentation Topology with configs
Updated the network - DHCP pools were created for all VLANs with exclusions for static addresses. The MLS was given a VLAN1 IP and default route, while Router0 was configured with a static route to the MLS and NAT for WAN testing. The attacker PC was able to get an IP but could not ping devices in other VLANs.